requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.4.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
python-dotenv==1.0.1
